create unique ids for primary keys

+++ app/src/main/java/com/carvajal/lucas/supertaster/data/CookbookRecipe.kt
@@ -13,7 +13,7 @@ import androidx.room.PrimaryKey
     ]
 )
 data class CookbookRecipe(
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "cookbook_id") val cookbookId: Long,
     @ColumnInfo(name = "recipe_id") val recipeId: Int,
 )
